# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API LatinAd",
        "description": "Documentación de la API de gestión de pantallas para LatinAd",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión de usuario",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "naranjaspintadas@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": 3600
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    }
                }
            }
        },
        "/api/displays": {
            "get": {
                "tags": [
                    "Displays"
                ],
                "summary": "Listar pantallas del usuario autenticado",
                "operationId": "ff135125e7af0e7cdd20a682525352a2",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtrar por nombre",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filtrar por tipo",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "indoor",
                                "outdoor"
                            ]
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Cantidad de resultados por página",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado paginado de pantallas"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Displays"
                ],
                "summary": "Crear nueva pantalla",
                "operationId": "9b1347658f33ccfa99770934d896cfea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "price_per_day",
                                    "resolution_height",
                                    "resolution_width",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Pantalla LED"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Pantalla ubicada en esquina comercial"
                                    },
                                    "price_per_day": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1000
                                    },
                                    "resolution_height": {
                                        "type": "integer",
                                        "example": 1080
                                    },
                                    "resolution_width": {
                                        "type": "integer",
                                        "example": 1920
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "indoor",
                                            "outdoor"
                                        ],
                                        "example": "indoor"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pantalla creada"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/displays/{id}": {
            "get": {
                "tags": [
                    "Displays"
                ],
                "summary": "Ver una pantalla por ID",
                "operationId": "2aad98a9c8739dbc56b65586d8e5b57e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pantalla encontrada"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Displays"
                ],
                "summary": "Actualizar pantalla existente",
                "operationId": "d2fce32a8fc35eddb9cde5c531412311",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "price_per_day",
                                    "resolution_height",
                                    "resolution_width",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Pantalla LED"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Actualizada con nueva ubicación"
                                    },
                                    "price_per_day": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 950
                                    },
                                    "resolution_height": {
                                        "type": "integer",
                                        "example": 720
                                    },
                                    "resolution_width": {
                                        "type": "integer",
                                        "example": 1280
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "indoor",
                                            "outdoor"
                                        ],
                                        "example": "outdoor"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pantalla actualizada"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Displays"
                ],
                "summary": "Eliminar una pantalla",
                "operationId": "fe332e375ec8835d8c3cc5c0c5a56838",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pantalla eliminada"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Autenticación",
            "description": "Autenticación"
        },
        {
            "name": "Displays",
            "description": "Displays"
        }
    ]
}